chore: moved mcp structs to config package
